refactor(twap): change window_seconds type from i64 to u64 in get_twaps_with_update_data function

+++ apps/hermes/server/src/api/rest/v2/latest_twaps.rs
@@ -109,7 +109,7 @@ where
     let twaps_with_update_data = Aggregates::get_twaps_with_update_data(
         &*state.state,
         &price_ids,
-        path_params.window_seconds as i64,
+        path_params.window_seconds,
         RequestTime::LatestTimeEarliestSlot,
     )
     .await
